Eval_llm/requirements.txt
litellm==1.61.12
python-dotenv==1.0.1
pandas==2.2.2
tqdm==4.66.5
